优化:将CommonTabLayout及SlidingTabLayout共有属性初始值进行调整;SegmentLayout初始属性进行调整 优化:将部分注释掉无用java方法删除;将原tab自带java方法增加返回本对象方便链式调用

+++ lib/src/main/java/com/aries/ui/view/tab/CommonTabLayout.java
@@ -371,8 +371,8 @@ public class CommonTabLayout extends FrameLayout implements ValueAnimator.Animat
                 }
 
                 mIndicatorDrawable.setColor(getDelegate().getIndicatorColor());
-                mIndicatorDrawable.setBounds(paddingLeft + (int) getDelegate().getIndicatorMarginLeft() + mIndicatorRect.left,
-                        (int) getDelegate().getIndicatorMarginTop(), (int) (paddingLeft + mIndicatorRect.right - getDelegate().getIndicatorMarginRight()),
+                mIndicatorDrawable.setBounds(paddingLeft + getDelegate().getIndicatorMarginLeft() + mIndicatorRect.left,
+                        getDelegate().getIndicatorMarginTop(), (paddingLeft + mIndicatorRect.right - getDelegate().getIndicatorMarginRight()),
                         (int) (getDelegate().getIndicatorMarginTop() + getDelegate().getIndicatorHeight()));
                 mIndicatorDrawable.setCornerRadius(getDelegate().getIndicatorCornerRadius());
                 mIndicatorDrawable.draw(canvas);
@@ -437,8 +437,6 @@ public class CommonTabLayout extends FrameLayout implements ValueAnimator.Animat
         TextView tv_tab_title = tabView.findViewById(R.id.tv_tab_title);
         return tv_tab_title;
     }
-
-    //setter and getter
 
     /**
      * show MsgTipView
